using hooks and update delete group function on the server"

+++ server/src/group/group.service.ts
@@ -5,7 +5,6 @@ import {
   forwardRef,
   Inject,
   Injectable,
-  InternalServerErrorException,
   NotFoundException,
 } from '@nestjs/common';
 import { Group } from './group.entity';
@@ -25,8 +24,8 @@ export class GroupService {
     @Inject(forwardRef(() => UserService))
     private readonly userService: UserService,
     @Inject(forwardRef(() => ExpenseService))
-    private readonly expenseService: ExpenseService
-  ) { }
+    private readonly expenseService: ExpenseService,
+  ) {}
 
   private toResponseDto(group: Group): FullGroupResponseDto {
     return plainToInstance(FullGroupResponseDto, group, {
@@ -113,7 +112,7 @@ export class GroupService {
   ): Promise<FullGroupResponseDto> {
     const group = await this.groupRepository.findOne({
       where: { uuid: groupId, owner: { uuid: userId } },
-      relations: ['owner','members','expenses'],
+      relations: ['owner', 'members', 'expenses'],
     });
     if (!group) {
       throw new NotFoundException('group not found');
@@ -152,13 +151,14 @@ export class GroupService {
     if (group.owner.uuid !== userId) {
       throw new ForbiddenException('Only the owner can delete this group');
     }
-    try {
-      await this.groupRepository.remove(group);
-    } catch {
-      throw new InternalServerErrorException(
-        'Could not delete group. Ensure all related data is cleared or cascading is enabled',
-      );
+    const groupBalancedExpense = await this.expenseService.getGroupExpense(
+      groupId,
+      userId,
+    );
+    if (groupBalancedExpense.length > 0) {
+      throw new BadRequestException('you cant delete unbalanced group');
     }
+    await this.groupRepository.remove(group);
     return { message: 'Group deleted successfully' };
   }
 
@@ -192,12 +192,17 @@ export class GroupService {
         'please assign someone else to be the owner of the group and request him to remove you',
       );
     }
-    const userExpenses = await this.expenseService.getGroupExpense(group.uuid, user.uuid)
-    const hasOpenDebts = userExpenses.some(expense =>
-      expense.paidByUser.uuid === user.uuid || expense.paidOnUser.uuid === user.uuid
-    )
+    const userExpenses = await this.expenseService.getGroupExpense(
+      group.uuid,
+      user.uuid,
+    );
+    const hasOpenDebts = userExpenses.some(
+      (expense) =>
+        expense.paidByUser.uuid === user.uuid ||
+        expense.paidOnUser.uuid === user.uuid,
+    );
     if (hasOpenDebts) {
-      throw new BadRequestException('user still has open expenses')
+      throw new BadRequestException('user still has open expenses');
     }
     const members = group?.members.filter(
       (member) => member.uuid !== userToRemoveId,
@@ -229,7 +234,7 @@ export class GroupService {
       throw new NotFoundException('user to add not found');
     }
     const isMember = group.members.some((member) => member.uuid == userToAddId);
-    
+
     if (isMember) {
       throw new ConflictException('user already in group');
     }
